A big change 😄 unified router and Secured parameters 😭 loginOut logout also changed to reload
