+++ src/components/Authorized/AuthorizedRoute.js
@@ -1,16 +1,14 @@
 import React from 'react';
 import { Route, Redirect } from 'dva/router';
 import Authorized from './Authorized';
-import { getAuthority } from '../../utils/authority';
 
 class AuthorizedRoute extends React.Component {
   render() {
-    const { component: Component, render, authority, currentAuthority = getAuthority(),
+    const { component: Component, render, authority,
       redirectPath, ...rest } = this.props;
     return (
       <Authorized
         authority={authority}
-        currentAuthority={currentAuthority}
         noMatch={<Route {...rest} render={() => <Redirect to={{ pathname: redirectPath }} />} />}
       >
         <Route
