+++ src/components/Authorized/index.js
@@ -1,8 +1,31 @@
 import Authorized from './Authorized';
 import AuthorizedRoute from './AuthorizedRoute';
-import RenderSecured from './Secured';
+import Secured from './Secured';
 
+/* eslint-disable import/no-mutable-exports */
+let CURRENT = 'NULL';
+
+Authorized.Secured = Secured;
 Authorized.AuthorizedRoute = AuthorizedRoute;
 
-export { RenderSecured };
-export default Authorized;
+
+/**
+ * use  authority or getAuthority
+ * @param {string|()=>String} currentAuthority
+ */
+const renderAuthorize = (currentAuthority) => {
+  if (currentAuthority) {
+    if (currentAuthority.constructor.name === 'Function') {
+      CURRENT = currentAuthority();
+    }
+    if (currentAuthority.constructor.name === 'String') {
+      CURRENT = currentAuthority;
+    }
+  } else {
+    CURRENT = 'NULL';
+  }
+  return Authorized;
+};
+
+export { CURRENT, Authorized };
+export default renderAuthorize;
