+++ src/router.js
@@ -4,11 +4,10 @@ import { LocaleProvider, Spin } from 'antd';
 import zhCN from 'antd/lib/locale-provider/zh_CN';
 import dynamic from 'dva/dynamic';
 import { getRouterData } from './common/router';
-import Authorized from './components/Authorized';
+import Authorized from './utils/Authorized';
 import styles from './index.less';
 
 const { AuthorizedRoute } = Authorized;
-
 dynamic.setDefaultLoadingComponent(() => {
   return <Spin size="large" className={styles.globalSpin} />;
 });
