+++ src/utils/Authorized.js
@@ -1,0 +1,5 @@
+import RenderAuthorized from '../components/Authorized';
+import { getAuthority } from './authority';
+
+const Authorized = RenderAuthorized(getAuthority());
+export default Authorized;
