--- src/utils/Secured.js
@@ -1,5 +1,0 @@
-import { RenderSecured } from '../components/Authorized';
-import { getAuthority } from './authority';
-
-const Secured = RenderSecured(getAuthority());
-export default Secured;
